Changes some file names.
